refactor: extract overlay logic from connection manager
