+++ InteractiveClassroom/Model/Interaction/InteractionHandling.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol InteractionHandling: AnyObject {
+    func startInteraction(_ request: InteractionRequest, broadcast: Bool)
+    func endInteraction(broadcast: Bool)
+}
